requirements.txt
openai == 1.93.0
python-dotenv == 1.1.1
fastapi == 0.115.14
uvicorn == 0.35.0
python-multipart == 0.0.20
pandas == 2.2.3
aiofiles == 24.1.0